feat(optical): show transfer speed + ETA in the CLI rip/convert progress
The GUI progress bar already showed rate + ETA; the CLI only printed percent.
Lift the rolling-window estimator into a shared, non-GUI module so both
surfaces use one implementation, and wire it into the CLI.

- model/rate_tracker.rs (new): RateTracker moved verbatim out of gui/progress.rs
  (record / rate_bytes_per_sec / eta_secs / suffix / reset) + its unit tests.
  Pure std, ungated — available to the CLI/mini build.
- gui/progress.rs: re-exports RateTracker from the model layer (the backup /
  restore / inspect / export tabs that use `progress::RateTracker` are unchanged).
- cli/verbs/optical.rs: drain_rip + drain_convert sample the tracker each tick
  and append " - <rate>/s, ETA <eta>" to the progress line, e.g.
  `  progress:  45% (315.0 MiB/700.0 MiB) - 28.4 MiB/s, ETA 13s`.
  For a remote rip the rate reflects LAN throughput; for CHD it's the local
  encode rate.

Builds + clippy clean (GUI + mini); rate_tracker tests green.
